added user module and payment module

+++ src/main/java/com/ngo/model/OrderItem.java
@@ -15,7 +15,7 @@ public class OrderItem {
     String productDescription;
     String productCategoryName;
     String productImageFile;
-    double quantity;
+    int quantity;
     double price;
     OrderItemStatus status;
     Date created;
